Scripting/csharp

+++ GloryEngine/EntityScenesEditorExtension/DefaultComponentEditor.cpp
@@ -27,7 +27,7 @@ namespace Glory::Editor
 		const GloryReflect::TypeData* pTypeData = GloryReflect::Reflect::GetTyeData(hash);
 		if (pTypeData)
 		{
-			for (size_t i = 0; i < pTypeData->FieldCount(); i++)
+			for (int i = 0; i < pTypeData->FieldCount(); i++)
 			{
 				const GloryReflect::FieldData* pFieldData = pTypeData->GetFieldData(i);
 				size_t offset = pFieldData->Offset();
