+++ GloryEngine/GloryEntityScenes/EntitySceneScenesModule.h
@@ -1,6 +1,5 @@
 #pragma once
 #include <ScenesModule.h>
-#include <MaterialData.h>
 #include <Glory.h>
 #include "EntityScene.h"
 #include "Entity.h"
@@ -26,7 +25,7 @@ namespace Glory
 
 		static EntitySceneObject* CreateDeserializedObject(GScene* pScene, const std::string& name, UUID uuid);
 
-		virtual SceneObject* GetSceneObjectFromObjectID(uint32_t objectID) override;
+		virtual SceneObject* GetSceneObjectFromObjectID(uint64_t objectID) override;
 
 	private:
 		friend class EntitySceneObjectSerializer;
